Normals fixed, shadow mapping class
Fixed the issue with normals being displayed incorrectly (was sending
normals as a vec2 instead of a vec3). Next step is to clean up the
Main.cpp file and put code into appropriate classes/functions. First
task is putting the current shadow mapping class into its own master
namespace/class so there can be both a sun source for light and
omnidirectional sources (which will be implemented in the future).